make basesprite base class
