+++ libs/game/pxt.json
@@ -3,6 +3,7 @@
     "description": "The game and sprite library - beta",
     "files": [
         "gameoverrides.ts",
+        "basesprite.ts",
         "constants.ts",
         "controlleroverrides.ts",
         "controller.ts",
